refactor: modularize express routes
